responsive styling

+++ src/components/ThreadsPage/ThreadsList/Thread/Thread.js
@@ -4,6 +4,8 @@ import styles from './Thread.module.scss';
 import Button from '../../../Button/Button';
 import ThreadsApiService from '../../../../services/threads-api-service';
 import ThreadsContext from '../../../../contexts/ThreadsContext';
+import Confirm from '../../../Confirm/Confirm';
+import '@reach/dialog/styles.css';
 
 export default class Thread extends React.Component {
   static contextType = ThreadsContext;
@@ -34,15 +36,18 @@ export default class Thread extends React.Component {
             Thread with {name}({user_name})
           </h3>
         </Link>
-        <button
-          type="delete"
-          className={styles.deleteThreadButton}
-          onClick={e => {
-            e.preventDefault();
-            this.handleThreadDelete();
-          }}>
-          <span className={styles.deleteX}>X</span>
-        </button>
+        <Confirm title="Confirm" description="Are you sure?">
+          {confirm => (
+            <button
+              type="delete"
+              className={styles.deleteThreadButton}
+              onClick={confirm(() => {
+                this.handleThreadDelete();
+              })}>
+              <span className={styles.deleteX}>X</span>
+            </button>
+          )}
+        </Confirm>
         <p className={styles.lastMessage}>
           last message: {thread.date_modified.slice(0, 10)}
         </p>
